+++ sdk/mobilenetwork/azure-mgmt-mobilenetwork/generated_samples/extended_ue_info5_gget.py
@@ -14,7 +14,7 @@ from azure.mgmt.mobilenetwork import MobileNetworkManagementClient
     pip install azure-identity
     pip install azure-mgmt-mobilenetwork
 # USAGE
-    python packet_core_control_plane_get.py
+    python extended_ue_info5_gget.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -29,13 +29,14 @@ def main():
         subscription_id="00000000-0000-0000-0000-000000000000",
     )
 
-    response = client.packet_core_control_planes.get(
+    response = client.extended_ue_information.get(
         resource_group_name="rg1",
         packet_core_control_plane_name="TestPacketCoreCP",
+        ue_id="84449105622",
     )
     print(response)
 
 
-# x-ms-original-file: specification/mobilenetwork/resource-manager/Microsoft.MobileNetwork/stable/2024-02-01/examples/PacketCoreControlPlaneGet.json
+# x-ms-original-file: specification/mobilenetwork/resource-manager/Microsoft.MobileNetwork/stable/2024-02-01/examples/ExtendedUeInfo5GGet.json
 if __name__ == "__main__":
     main()